Extract the pages menu
